Rename `validatePeersDeps` to `validatePeerDeps`
