+++ lib/validatePeerDeps.js
